feat(PracticeQuestion): CRUD PracticeQuestion full

+++ BackEnd/LIP.Application/CQRS/Handler/Practicequestion/PracticequestionDeleteCommandHandler.cs
@@ -1,10 +1,12 @@
 using LIP.Application.CQRS.Command.Practicequestion;
+using LIP.Application.CQRS.Query.Practicequestion;
+using LIP.Application.DTOs.Response.PracticeQuestion;
 using LIP.Application.Interface.Repository;
 using MediatR;
 
 namespace LIP.Application.CQRS.Handler.Practicequestion
 {
-    public class PracticequestionDeleteCommandHandler : IRequestHandler<PracticequestionDeleteCommand, bool>
+    public class PracticequestionDeleteCommandHandler : IRequestHandler<PracticequestionDeleteCommand, PracticeQuestionDeleteResponse>
     {
         private readonly IPracticequestionRepository _practicequestionRepository;
 
@@ -13,9 +15,39 @@ namespace LIP.Application.CQRS.Handler.Practicequestion
             _practicequestionRepository = practicequestionRepository;
         }
 
-        public async Task<bool> Handle(PracticequestionDeleteCommand request, CancellationToken cancellationToken)
+        public async Task<PracticeQuestionDeleteResponse> Handle(PracticequestionDeleteCommand request, CancellationToken cancellationToken)
         {
-            return await _practicequestionRepository.DeleteAsync(request);
+            var question = await _practicequestionRepository.GetAsync(new PracticequestionGetQuery { QuestionId = request.QuestionId });
+            if (question == null)
+            {
+                return new PracticeQuestionDeleteResponse
+                {
+                    IsSuccess = false,
+                    Message = "Practice Question is not found!"
+                };
+            }
+            var rs = await _practicequestionRepository.DeleteAsync(request);
+            if (rs)
+            {
+                return new PracticeQuestionDeleteResponse
+                {
+                    IsSuccess = true,
+                    Message = "Delete Practice Question successfully!"
+                };
+            }
+            else
+            {
+                return new PracticeQuestionDeleteResponse
+                {
+                    IsSuccess = false,
+                    Message = "Delete Practice Question failed!"
+                };
+            }
         }
+
+        //public async Task<bool> Handle(PracticequestionDeleteCommand request, CancellationToken cancellationToken)
+        //{
+        //    return await _practicequestionRepository.DeleteAsync(request);
+        //}
     }
 }
